+++ src/TrashLib/Sonarr/IReleaseProfileLister.cs
@@ -3,4 +3,5 @@ namespace TrashLib.Sonarr;
 public interface IReleaseProfileLister
 {
     void ListReleaseProfiles();
+    void ListTerms(string releaseProfileId);
 }