+++ src/TrashLib/Sonarr/ReleaseProfile/Guide/ISonarrGuideService.cs
@@ -3,4 +3,5 @@
 public interface ISonarrGuideService
 {
     IReadOnlyCollection<ReleaseProfileData> GetReleaseProfileData();
+    ReleaseProfileData? GetUnfilteredProfileById(string trashId);
 }