+++ src/TrashLib/Sonarr/ReleaseProfile/Guide/LocalRepoReleaseProfileJsonParser.cs
@@ -1,4 +1,5 @@
 using System.IO.Abstractions;
+using Common.Extensions;
 using Common.FluentValidation;
 using MoreLinq;
 using Newtonsoft.Json;
@@ -10,14 +11,16 @@ public class LocalRepoReleaseProfileJsonParser : ISonarrGuideService
 {
     private readonly IFileSystem _fileSystem;
     private readonly IResourcePaths _paths;
+    private readonly Lazy<IEnumerable<ReleaseProfileData>> _data;
 
     public LocalRepoReleaseProfileJsonParser(IFileSystem fileSystem, IResourcePaths paths)
     {
         _fileSystem = fileSystem;
         _paths = paths;
+        _data = new Lazy<IEnumerable<ReleaseProfileData>>(GetReleaseProfileDataImpl);
     }
 
-    public IReadOnlyCollection<ReleaseProfileData> GetReleaseProfileData()
+    private IEnumerable<ReleaseProfileData> GetReleaseProfileDataImpl()
     {
         var converter = new TermDataConverter();
         var jsonDir = Path.Combine(_paths.RepoPath, "docs/json/sonarr");
@@ -29,7 +32,17 @@ public class LocalRepoReleaseProfileJsonParser : ISonarrGuideService
             });
 
         return Task.WhenAll(tasks).Result
-            .Choose(x => x is not null ? (true, x) : default) // Make non-nullable type
+            .Choose(x => x is not null ? (true, x) : default); // Make non-nullable type
+    }
+
+    public ReleaseProfileData? GetUnfilteredProfileById(string trashId)
+    {
+        return _data.Value.FirstOrDefault(x => x.TrashId.EqualsIgnoreCase(trashId));
+    }
+
+    public IReadOnlyCollection<ReleaseProfileData> GetReleaseProfileData()
+    {
+        return _data.Value
             .IsValid(new ReleaseProfileDataValidator())
             .ToList();
     }