+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.0.0] - 2022-05-13
+
 This release contains **BREAKING CHANGES**. See the [Upgrade Guide] for required changes you need to
 make.
 
@@ -266,7 +268,8 @@ See the [Python Migration Guide][py-mig] for details on how to update your YAML 
 - Nearly all command line options removed in favor of YAML equivalents.
 - Completely removed old python project & source code
 
-[Unreleased]: https://github.com/rcdailey/recyclarr/compare/v1.8.2...HEAD
+[Unreleased]: https://github.com/rcdailey/recyclarr/compare/v2.0.0...HEAD
+[2.0.0]: https://github.com/rcdailey/recyclarr/compare/v1.8.2...v2.0.0
 [1.8.2]: https://github.com/rcdailey/recyclarr/compare/v1.8.1...v1.8.2
 [1.8.1]: https://github.com/rcdailey/recyclarr/compare/v1.8.0...v1.8.1
 [1.8.0]: https://github.com/rcdailey/recyclarr/compare/v1.7.0...v1.8.0